# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Evolvus Swagger API",
    "description": "Evolvus Spec for Swagger",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/api/entity/{entityCode}": {
      "put": {
        "tags": [
          "Entity API"
        ],
        "summary": "Update an entity to the database",
        "description": "",
        "operationId": "updateEntity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Entity object that needs to be updated to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entity"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/entity": {
      "post": {
        "tags": [
          "Entity API"
        ],
        "summary": "Add a new entity to the database",
        "description": "",
        "operationId": "addEntity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Entity object that needs to be added to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entity"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "Entity API"
        ],
        "summary": "Get an entity from the database.",
        "description": "",
        "operationId": "getEntity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Entity object that needs to be get from the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Entity"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "api/application": {
      "get": {
        "tags": [
          "Application API"
        ],
        "summary": "Get an application from the database.",
        "description": "",
        "operationId": "getApplication",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "Application API"
        ],
        "summary": "Add a new application to the database",
        "description": "",
        "operationId": "addApplication",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Application object that needs to be added to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Application"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/application/{applicationCode}": {
      "put": {
        "tags": [
          "Application API"
        ],
        "summary": "Update an application to the database",
        "description": "",
        "operationId": "updateApplication",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Application object that needs to be updated to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Entity": {
      "type": "object",
      "required": [
        "name",
        "entityCode",
        "tenantId",
        "parent"
      ],
      "tags": [
        {
          "name": "string",
          "entityCode": "string",
          "description": "string",
          "tenantId": "string",
          "parent": "string"
        }
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "MGRoad"
        },
        "entityCode": {
          "type": "string",
          "example": "MGRoad"
        },
        "description": {
          "type": "string",
          "example": "This is MGRoad Entity"
        },
        "tenantId": {
          "type": "string",
          "example": "T001"
        },
        "parent": {
          "type": "string",
          "example": "HEADOFFICET002"
        }
      }
    },
...
    "Application": {
      "type": "object",
      "required": [
        "tenantId",
        "applicationName",
        "applicationCode",
        "entityId"
      ],
      "tags": [
        {
          "tenantId": "string",
          "entityId": "string",
          "description": "string",
          "applicationName": "string",
          "applicationCode": "string",
          "logo": "string",
          "favicon": "string"
        }
      ],
      "properties": {
        "tenantId": {
          "type": "string",
          "example": "T001"
        },
        "applicationName": {
          "type": "string",
          "example": "SANDSTORM"
        },
        "applicationCode": {
          "type": "string",
          "example": "SANDSTORM"
        },
        "entityId": {
          "type": "string",
          "example": "H001B001"
        },
        "description": {
          "type": "string",
          "example": "This is application Sandstorm"
        }
      }
    },
...
  }
}